+++ wear/src/main/java/io/homeassistant/companion/android/onboarding/OnboardingPresenter.kt
@@ -1,8 +1,13 @@
 package io.homeassistant.companion.android.onboarding
 
-interface OnboardingPresenter {
+import com.google.android.gms.wearable.DataClient
+import com.google.android.gms.wearable.DataMap
+
+interface OnboardingPresenter: DataClient.OnDataChangedListener {
 
     fun onAdapterItemClick(instance: HomeAssistantInstance)
 
     fun onFinish()
+
+    fun getInstance(map: DataMap): HomeAssistantInstance
 }
