+++ wear/src/main/java/io/homeassistant/companion/android/onboarding/OnboardingPresenterImpl.kt
@@ -1,12 +1,15 @@
 package io.homeassistant.companion.android.onboarding
 
 import android.util.Log
+import com.google.android.gms.wearable.DataEvent
+import com.google.android.gms.wearable.DataEventBuffer
+import com.google.android.gms.wearable.DataMap
+import com.google.android.gms.wearable.DataMapItem
 import io.homeassistant.companion.android.common.data.authentication.AuthenticationRepository
-import io.homeassistant.companion.android.common.data.authentication.impl.entities.LoginFlowCreateEntry
 import io.homeassistant.companion.android.common.data.authentication.impl.entities.LoginFlowInit
 import io.homeassistant.companion.android.common.data.url.UrlRepository
-import io.homeassistant.companion.android.onboarding.authentication.AuthenticationActivity
 import kotlinx.coroutines.*
+import java.net.URL
 import javax.inject.Inject
 
 class OnboardingPresenterImpl @Inject constructor(
@@ -19,6 +22,38 @@ class OnboardingPresenterImpl @Inject constructor(
     }
 
     private val mainScope: CoroutineScope = CoroutineScope(Dispatchers.Main + Job())
+
+    override fun onDataChanged(dataEvents: DataEventBuffer) {
+        Log.d(TAG, "onDataChanged: [${dataEvents.count}]")
+        dataEvents.forEach { event ->
+            if (event.type == DataEvent.TYPE_CHANGED) {
+                event.dataItem.also { item ->
+                    if (item.uri.path?.compareTo("/home_assistant_instance") == 0) {
+                        Log.d(TAG, "onDataChanged: found home_assistant_instance")
+                        val instance = getInstance(DataMapItem.fromDataItem(item).dataMap)
+                        view.onInstanceFound(instance)
+                    }
+                }
+            } else if (event.type == DataEvent.TYPE_DELETED) {
+                event.dataItem.also { item ->
+                    if (item.uri.path?.compareTo("/home_assistant_instance") == 0) {
+                        val instance = getInstance(DataMapItem.fromDataItem(item).dataMap)
+                        view.onInstanceLost(instance)
+                    }
+                }
+            }
+        }
+    }
+
+    override fun getInstance(map: DataMap): HomeAssistantInstance {
+        map.apply {
+            return HomeAssistantInstance(
+                getString("name", ""),
+                URL(getString("url", "")),
+                getString("version", "")
+            )
+        }
+    }
 
     override fun onAdapterItemClick(instance: HomeAssistantInstance) {
         Log.d(TAG, "onAdapterItemClick: ${instance.name}")
