+++ wear/src/main/java/io/homeassistant/companion/android/onboarding/OnboardingView.kt
@@ -4,6 +4,9 @@ interface OnboardingView {
     fun startAuthentication(flowId: String)
     fun startManualSetup()
 
+    fun onInstanceFound(instance: HomeAssistantInstance)
+    fun onInstanceLost(instance: HomeAssistantInstance)
+
     fun showLoading()
 
     fun showError()
