+++ wear/src/main/java/io/homeassistant/companion/android/onboarding/ServerListAdapter.kt
@@ -1,6 +1,5 @@
 package io.homeassistant.companion.android.onboarding
 
-import android.content.res.Resources
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
