+++ wear/src/main/java/io/homeassistant/companion/android/onboarding/viewHolders/HeaderViewHolder.kt
@@ -1,6 +1,5 @@
 package io.homeassistant.companion.android.onboarding.viewHolders
 
-import android.util.Log
 import android.view.View
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
