+++ wear/src/main/java/io/homeassistant/companion/android/onboarding/viewHolders/LoadingViewHolder.kt
@@ -1,10 +1,7 @@
 package io.homeassistant.companion.android.onboarding.viewHolders
 
-import android.util.Log
 import android.view.View
-import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import io.homeassistant.companion.android.R
 
 class LoadingViewHolder(v: View): RecyclerView.ViewHolder(v) {
 }
